[arcane] Déplace certaines classes et interfaces de 'ceapart/src/arcane/materials' vers 'arcane/core'
