+++ arcane/src/arcane/core/materials/MaterialsCoreGlobal.cc
@@ -1,0 +1,21 @@
+﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
+//-----------------------------------------------------------------------------
+// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: Apache-2.0
+//-----------------------------------------------------------------------------
+/*---------------------------------------------------------------------------*/
+/* MaterialsCoreGlobal.cc                                      (C) 2000-2022 */
+/*                                                                           */
+/* Déclarations générales des matériaux de Arcane.                           */
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+#include "arcane/core/materials/IMeshMaterialVariable.h"
+#include "arcane/core/materials/IMeshBlock.h"
+#include "arcane/core/materials/IMeshComponent.h"
+#include "arcane/core/materials/IMeshMaterial.h"
+#include "arcane/core/materials/IMeshEnvironment.h"
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
