+++ arcane/src/arcane/core/materials/MeshMaterialVariableIndexer.cc
@@ -5,7 +5,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* MeshMaterialVariableIndexer.cc                              (C) 2000-2016 */
+/* MeshMaterialVariableIndexer.cc                              (C) 2000-2022 */
 /*                                                                           */
 /* Indexer pour les variables materiaux.                                     */
 /*---------------------------------------------------------------------------*/
@@ -17,20 +17,14 @@
 #include "arcane/utils/ArgumentException.h"
 #include "arcane/utils/NotSupportedException.h"
 
-#include "arcane/materials/MeshMaterialVariableIndexer.h"
-#include "arcane/materials/ComponentItemListBuilder.h"
+#include "arcane/core/materials/MeshMaterialVariableIndexer.h"
+#include "arcane/core/materials/ComponentItemListBuilder.h"
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-ARCANE_BEGIN_NAMESPACE
-MATERIALS_BEGIN_NAMESPACE
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
+namespace Arcane::Materials
+{
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -380,8 +374,7 @@ checkValid()
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-MATERIALS_END_NAMESPACE
-ARCANE_END_NAMESPACE
+} // End namespace Arcane::Materials
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
